requirements.txt
gradio
transformers @ git+https://github.com/huggingface/transformers
torch
torchvision
opencv-python-headless
ffmpeg-python
tqdm
pillow
supervision
trackers[deepsort] @ git+https://github.com/roboflow/trackers
spaces
imageio[pyav]
boxmot